+++ src/chat.rs
@@ -4693,7 +4693,9 @@ mod tests {
         assert!(!shall_attach_selfavatar(&t, chat_id).await?);
 
         let (contact_id, _) =
-            Contact::add_or_lookup(&t, "", "foo@bar.org", Origin::IncomingUnknownTo).await?;
+            Contact::add_or_lookup(&t, "", "foo@bar.org", Origin::IncomingUnknownTo)
+                .await?
+                .unwrap();
         add_contact_to_chat(&t, chat_id, contact_id).await?;
         assert!(!shall_attach_selfavatar(&t, chat_id).await?);
         t.set_config(Config::Selfavatar, None).await?; // setting to None also forces re-sending
@@ -4940,7 +4942,9 @@ mod tests {
         bob.set_config(Config::ShowEmails, Some("2")).await?;
 
         let (contact_id, _) =
-            Contact::add_or_lookup(&alice, "", "bob@example.net", Origin::ManuallyCreated).await?;
+            Contact::add_or_lookup(&alice, "", "bob@example.net", Origin::ManuallyCreated)
+                .await?
+                .unwrap();
         let alice_chat_id = create_group_chat(&alice, ProtectionStatus::Unprotected, "grp").await?;
         let alice_chat = Chat::load_from_db(&alice, alice_chat_id).await?;
 
@@ -5651,7 +5655,9 @@ mod tests {
     async fn test_create_for_contact_with_blocked() -> Result<()> {
         let t = TestContext::new().await;
         let (contact_id, _) =
-            Contact::add_or_lookup(&t, "", "foo@bar.org", Origin::ManuallyCreated).await?;
+            Contact::add_or_lookup(&t, "", "foo@bar.org", Origin::ManuallyCreated)
+                .await?
+                .unwrap();
 
         // create a blocked chat
         let chat_id_orig =